// resources/js/entities/vehicle/model/types.ts
export interface Vehicle {
    id: number;
    make: string;
    model: string;
    title: string;
    description: string;
    price: number;
    mileage: number;
    fuel_type: string;
    power: number;
    transmission: string;
    drive: string;
    year: number;
    seller: string;
    city: string;
    date: string;
    image: string;
    verified: boolean;
    tag?: string;
}